Andreasen 2012: rare disaster model with asymmetric shocks
Replicates Table 2 in Andreasen (2012): "On the effects of rare disasters and uncertainty shocks for risk premia in non-linear DSGE models", Review of Economic Dynamics, 15, pp. 295-316.

This mod file shows how to simulate DSGE models solved with third-order perturbation and non-symmetric innovations.
The underlying model is a New-Keynesian model with Epstein-Zin preferences.